requirements.txt
mne
PyYAML
tabulate
prettytable
numpy
matplotlib